qemu-tests: More coverage
All that is missing now for 100 % coverage are tests that require sudo
(well, they are covered, if you have that), and weird protocols (nfs,
ssh) which would require someone to set up the server manually.

We need a small hack: iotests-rare runs two qcow2 tests in the main
iotests directory because the usual $__fname directory would be too
long for 199 (which creates an AF_UNIX socket in its scratch directory;
and for this test, the scratch directory is in the test directory so we
can use -c none -- anyway, AF_UNIX really dislikes long pathnames).
This should be fine as long as its the only target doing this.