Allow failure
Continue execution even after terminal failure.  All depending jobs are
skipped.

Also, add an "after" key to the job description that specifies soft
dependencies; for these, it does not matter whether they succeed or
fail.  Once they are removed from the job pool, the dependency counts as
fulfilled.
This is used by the fedora/all job of qemu-tests, so that the temporary
directory is always removed after the tests are run, even on failure.
And for good measure, let's make all pseudo-jobs use "after" instead of
"depends".